+++ src/main/scala/org/apache/spark/executor/CoarseGrainedExecutorBackend.scala
@@ -34,7 +34,7 @@ import org.apache.spark.rpc._
 import org.apache.spark.scheduler.{ExecutorLossReason, TaskDescription}
 import org.apache.spark.scheduler.cluster.CoarseGrainedClusterMessages._
 import org.apache.spark.serializer.SerializerInstance
-import org.apache.spark.sql.execution.datasources.oap.filecache.FiberCacheManager
+import org.apache.spark.sql.execution.datasources.oap.{OapMessage, OapMessageUtils}
 import org.apache.spark.util.{ThreadUtils, Utils}
 
 private[spark] class CoarseGrainedExecutorBackend(
@@ -125,8 +125,8 @@ private[spark] class CoarseGrainedExecutorBackend(
         }
       }.start()
 
-    case CacheDrop(indexName) =>
-      FiberCacheManager.removeIndexCache(indexName)
+    case message: OapMessage =>
+      OapMessageUtils.handleOapMessage(message)
   }
 
   override def onDisconnected(remoteAddress: RpcAddress): Unit = {
